requirements.txt
spacy
grpcio
hvac
google-api-python-client
elasticsearch==6.8.0
boto3